requirements.txt
statsbombpy==1.14.0
python-dotenv==1.0.1
google-generativeai==0.8.3
streamlit==1.41.1
fastapi==0.115.6